test: separate Fixture
