+++ src/test/java/kitchenpos/application/OrderTableServiceTest.java
@@ -17,7 +17,7 @@ import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
-import static kitchenpos.fixture.TestFixture.TEST_ORDER_TABLE;
+import static kitchenpos.fixture.OrderTableFixture.TEST_ORDER_TABLE;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.any;
