+++ src/test/java/kitchenpos/fixture/MenuFixture.java
@@ -1,0 +1,38 @@
+package kitchenpos.fixture;
+
+import kitchenpos.domain.Menu;
+import kitchenpos.domain.MenuGroup;
+import kitchenpos.domain.MenuProduct;
+import kitchenpos.domain.Product;
+
+import java.math.BigDecimal;
+import java.util.List;
+import java.util.UUID;
+
+import static kitchenpos.fixture.MenuGroupFixture.TEST_MENU_GROUP;
+import static kitchenpos.fixture.ProductFixture.TEST_PRODUCT;
+
+public class MenuFixture {
+
+    public static Menu TEST_MENU() {
+        Menu menu = new Menu();
+        menu.setName("테스트 메뉴");
+        menu.setPrice(new BigDecimal(10_00));
+        MenuGroup menuGroup = TEST_MENU_GROUP();
+        menu.setMenuGroup(menuGroup);
+        menu.setDisplayed(true);
+        menu.setMenuProducts(List.of(TEST_MENU_PRODUCT()));
+        menu.setMenuGroupId(menuGroup.getId());
+        menu.setId(UUID.randomUUID());
+        return menu;
+    }
+
+    public static MenuProduct TEST_MENU_PRODUCT() {
+        MenuProduct menuProduct = new MenuProduct();
+        menuProduct.setQuantity(3);
+        Product product = TEST_PRODUCT();
+        menuProduct.setProductId(product.getId());
+        menuProduct.setProduct(product);
+        return menuProduct;
+    }
+}
